Update collective framework count/disp arrays for bigcount

+++ ompi/mca/coll/portals4/coll_portals4_scatter.c
@@ -306,8 +306,8 @@ err_hdlr:
 }
 
 static int
-ompi_coll_portals4_scatter_intra_linear_top(const void *sbuf, int scount, struct ompi_datatype_t *sdtype,
-                                            void *rbuf, int rcount, struct ompi_datatype_t *rdtype,
+ompi_coll_portals4_scatter_intra_linear_top(const void *sbuf, size_t scount, struct ompi_datatype_t *sdtype,
+                                            void *rbuf, size_t rcount, struct ompi_datatype_t *rdtype,
                                             int root,
                                             struct ompi_communicator_t *comm,
                                             ompi_coll_portals4_request_t *request,
@@ -646,8 +646,8 @@ err_hdlr:
 }
 
 int
-ompi_coll_portals4_scatter_intra(const void *sbuf, int scount, struct ompi_datatype_t *sdtype,
-                                 void *rbuf, int rcount, struct ompi_datatype_t *rdtype,
+ompi_coll_portals4_scatter_intra(const void *sbuf, size_t scount, struct ompi_datatype_t *sdtype,
+                                 void *rbuf, size_t rcount, struct ompi_datatype_t *rdtype,
                                  int root,
                                  struct ompi_communicator_t *comm,
                                  mca_coll_base_module_t *module)
@@ -705,8 +705,8 @@ err_hdlr:
 
 
 int
-ompi_coll_portals4_iscatter_intra(const void *sbuf, int scount, struct ompi_datatype_t *sdtype,
-                                 void *rbuf, int rcount, struct ompi_datatype_t *rdtype,
+ompi_coll_portals4_iscatter_intra(const void *sbuf, size_t scount, struct ompi_datatype_t *sdtype,
+                                 void *rbuf, size_t rcount, struct ompi_datatype_t *rdtype,
                                  int root,
                                  struct ompi_communicator_t *comm,
                                  ompi_request_t **ompi_request,
